--- core/auth/logout/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path
-from .views import LogoutView
-
-urlpatterns = [
-    path(
-        "auth/logout/",
-        LogoutView.as_view(),
-        name="logout",
-    ),
-]
