--- core/auth/refresh/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path
-from .views import TokenRefreshView
-
-urlpatterns = [
-    path(
-        "auth/refresh/",
-        TokenRefreshView.as_view(),
-        name="token_refresh",
-    ),
-]
