--- core/study/create/urls.py
@@ -1,10 +1,0 @@
-from django.urls import path
-from .views import StudyCreateView
-
-urlpatterns = [
-    path(
-        "studies/",
-        StudyCreateView.as_view(),
-        name="study_create",
-    ),
-]
