--- core/user/me/views.py
@@ -1,14 +1,0 @@
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework import status
-from rest_framework.permissions import IsAuthenticated
-
-from .serializers import UserInfoSerializer
-
-
-class UserMeView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        serializer = UserInfoSerializer(request.user)
-        return Response(serializer.data, status=status.HTTP_200_OK)
