+++ core/user/urls.py
@@ -1,10 +1,22 @@
 from django.urls import path
 from .views import (
+    UserMeView,
+    RegisterProfileView,
     UsernameValidationView,
     BojUsernameValidationView,
 )
 
 urlpatterns = [
+    path(
+        "user/me/",
+        UserMeView.as_view(),
+        name="user_me",
+    ),
+    path(
+        "user/profile/",
+        RegisterProfileView.as_view(),
+        name="register_profile",
+    ),
     path(
         "user/validate/username/",
         UsernameValidationView.as_view(),
